requirements.txt
fastapi
uvicorn[standard]
python-multipart
easyocr
Pillow
fpdf2
